backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
anyio==4.7.0
bcrypt==4.2.1
